Set DeviceGray as initial value for color space
See the PDF reference, section 4.3 Graphics State, table 4.2, third row.

> The current color space in which color values are to be interpreted
> (see Section 4.5, “Color Spaces”). There are two separate color space
> parameters: one for stroking and one for all other painting opera-
> tions. Initial value: DeviceGray.

The problem before was that certain PDFs didn't explicitly set the
color space, so a call to `setFillColor` or `setStrokeColor` were failing
when the `getRgb` method was call, as the color space was null.

See source code of CanvasGraphics.prototype.setFillColor/setStrokeColor.